requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
coverage==6.3.1
Django==4.0.2
django-countries==7.2.1
django-debug-toolbar==3.2.4
django-js-asset==2.0.0
django-mptt==0.13.4
flake8==4.0.1
idna==3.3
mccabe==0.6.1
Pillow==9.0.1
pycodestyle==2.8.0
pyflakes==2.4.0
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
stripe==2.65.0
stripe3==0.0.0
tzdata==2021.5
urllib3==1.26.8
six==1.16.0
numpy==1.22.2
paypal-checkout-serversdk==1.0.1
paypalhttp==1.0.0


